Split AprilTagTrackers.cpp into MyApp.cpp and Tracker.cpp
